// student-growth-system-design-v3/student-growth-system-design-v3/frontend/src/api/generated/models/ExamType.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Student Growth Archive API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 3.15.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const ExamType = {
    DailyTest: 'DAILY_TEST',
    UnitTest: 'UNIT_TEST',
    WeeklyTest: 'WEEKLY_TEST',
    MonthlyExam: 'MONTHLY_EXAM',
    Midterm: 'MIDTERM',
    Final: 'FINAL',
    MockExam: 'MOCK_EXAM',
    Competition: 'COMPETITION',
    Other: 'OTHER'
} as const;
export type ExamType = typeof ExamType[keyof typeof ExamType];


export function instanceOfExamType(value: any): boolean {
    for (const key in ExamType) {
        if (Object.prototype.hasOwnProperty.call(ExamType, key)) {
            if (ExamType[key as keyof typeof ExamType] === value) {
                return true;
            }
        }
    }
    return false;
}

export function ExamTypeFromJSON(json: any): ExamType {
    return ExamTypeFromJSONTyped(json, false);
}

export function ExamTypeFromJSONTyped(json: any, ignoreDiscriminator: boolean): ExamType {
    return json as ExamType;
}

export function ExamTypeToJSON(value?: ExamType | null): any {
    return value as any;
}

export function ExamTypeToJSONTyped(value: any, ignoreDiscriminator: boolean): ExamType {
    return value as ExamType;
}